Don't add queueworker unless async is set to true
Currently queueworker is set to be started even if async is set to false. Since it doesn't seem to be needed unless async is set to true, and there is no arm image for queueworker, it is only to be installed if async=true. This also fixes the paths to the charts in chart/openfaas/README.md

Signed-off-by: David Muckle <dvdmuckle@dvdmuckle.xyz>